Bug fix: App no longer crashes when going in to dark mode inside Sensors Fragment
Problem: App was crashing  when going in to dark mode inside Sensors Fragment.
Problem: Gravity sensor Event Listener was not un registered when the fragment was destroyed.
Fix: When then fragment is closed first the sensor is unregistered.
Improvement: Now the sensor is unregistered when the fragment is close by any methods, so it doesn't have to be manually turned off now.